Split EmbedPython struct to own module.
